requirements.txt
flask==2.3.3
Werkzeug==2.3.7
pypdf==3.17.0
PyPDF2==3.0.1
python-docx==1.0.1
nltk==3.8.1
spacy==3.7.2
Pillow==10.2.0
pytesseract==0.3.10
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
regex==2023.10.3
python-dotenv==1.0.0
Flask-WTF==1.2.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
gunicorn==21.2.0 